requirements.txt
asgiref==3.4.1
Django==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
pymongo==3.12.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.2.4
